coinex: convert to TS
